[NEW] Big update to support settings and multiple target languages
This implements a first version of settings and multiple target languages. This was requested a few times. To implement this, I had to rewrite the whole thing almost completely ... again ;-)
The settings are not perfect yet. They write the data as soon as some valid input is typed in. I think it might me better to write the changes on enter. But to achieve that, a little more work has to be done. Anyways, this is quite usable already.

Some documentation is missing, too.